Move MLX audio models to dedicated module
